requirements.txt
wheel
flask
numpy
psycopg2-binary
psycopg2
python-dotenv